+++ pkg/bindinfo/global_handle.go
@@ -19,15 +19,11 @@ import (
 	"strings"
 	"time"
 
-	"github.com/pingcap/errors"
-	"github.com/pingcap/failpoint"
 	"github.com/pingcap/tidb/pkg/parser/ast"
 	"github.com/pingcap/tidb/pkg/parser/format"
-	"github.com/pingcap/tidb/pkg/parser/mysql"
 	"github.com/pingcap/tidb/pkg/sessionctx"
 	"github.com/pingcap/tidb/pkg/sessionctx/vardef"
 	"github.com/pingcap/tidb/pkg/sessionctx/variable"
-	"github.com/pingcap/tidb/pkg/types"
 	"github.com/pingcap/tidb/pkg/util"
 	"github.com/pingcap/tidb/pkg/util/chunk"
 	"github.com/pingcap/tidb/pkg/util/hint"
@@ -38,18 +34,7 @@ import (
 type GlobalBindingHandle interface {
 	BindingCacheUpdater
 
-	// CreateGlobalBinding creates a Bindings to the storage and the cache.
-	// It replaces all the exists bindings for the same normalized SQL.
-	CreateGlobalBinding(sctx sessionctx.Context, bindings []*Binding) (err error)
-
-	// DropGlobalBinding drop Bindings to the storage and Bindings int the cache.
-	DropGlobalBinding(sqlDigests []string) (deletedRows uint64, err error)
-
-	// SetGlobalBindingStatus set a Bindings's status to the storage and bind cache.
-	SetGlobalBindingStatus(newStatus, sqlDigest string) (ok bool, err error)
-
-	// GCGlobalBinding physically removes the deleted bind records in mysql.bind_info.
-	GCGlobalBinding() (err error)
+	BindingOperator
 
 	variable.Statistics
 }
@@ -57,7 +42,7 @@ type GlobalBindingHandle interface {
 // globalBindingHandle is used to handle all global sql bind operations.
 type globalBindingHandle struct {
 	BindingCacheUpdater
-	sPool util.DestroyableSessionPool
+	BindingOperator
 }
 
 // Lease influences the duration of loading bind info and handling invalid bind.
@@ -86,175 +71,10 @@ const (
 // NewGlobalBindingHandle creates a new GlobalBindingHandle.
 func NewGlobalBindingHandle(sPool util.DestroyableSessionPool) GlobalBindingHandle {
 	cache := NewBindingCacheUpdater(sPool)
-	h := &globalBindingHandle{sPool: sPool, BindingCacheUpdater: cache}
+	op := newBindingOperator(sPool, cache)
+	h := &globalBindingHandle{BindingOperator: op, BindingCacheUpdater: cache}
 	variable.RegisterStatistics(h)
 	return h
-}
-
-// CreateGlobalBinding creates a Bindings to the storage and the cache.
-// It replaces all the exists bindings for the same normalized SQL.
-func (h *globalBindingHandle) CreateGlobalBinding(sctx sessionctx.Context, bindings []*Binding) (err error) {
-	for _, binding := range bindings {
-		if err := prepareHints(sctx, binding); err != nil {
-			return err
-		}
-	}
-	defer func() {
-		if err == nil {
-			err = h.LoadFromStorageToCache(false)
-		}
-	}()
-
-	return callWithSCtx(h.sPool, true, func(sctx sessionctx.Context) error {
-		// Lock mysql.bind_info to synchronize with CreateBinding / AddBinding / DropBinding on other tidb instances.
-		if err = lockBindInfoTable(sctx); err != nil {
-			return err
-		}
-
-		for i, binding := range bindings {
-			now := types.NewTime(types.FromGoTime(time.Now()), mysql.TypeTimestamp, 3)
-
-			updateTs := now.String()
-			_, err = exec(
-				sctx,
-				`UPDATE mysql.bind_info SET status = %?, update_time = %? WHERE original_sql = %? AND update_time < %?`,
-				StatusDeleted,
-				updateTs,
-				binding.OriginalSQL,
-				updateTs,
-			)
-			if err != nil {
-				return err
-			}
-
-			binding.CreateTime = now
-			binding.UpdateTime = now
-
-			// Insert the Bindings to the storage.
-			_, err = exec(
-				sctx,
-				`INSERT INTO mysql.bind_info VALUES (%?,%?, %?, %?, %?, %?, %?, %?, %?, %?, %?)`,
-				binding.OriginalSQL,
-				binding.BindSQL,
-				strings.ToLower(binding.Db),
-				binding.Status,
-				binding.CreateTime.String(),
-				binding.UpdateTime.String(),
-				binding.Charset,
-				binding.Collation,
-				binding.Source,
-				binding.SQLDigest,
-				binding.PlanDigest,
-			)
-			failpoint.Inject("CreateGlobalBindingNthFail", func(val failpoint.Value) {
-				n := val.(int)
-				if n == i {
-					err = errors.NewNoStackError("An injected error")
-				}
-			})
-			if err != nil {
-				return err
-			}
-		}
-		return nil
-	})
-}
-
-// DropGlobalBinding drop Bindings to the storage and Bindings int the cache.
-func (h *globalBindingHandle) DropGlobalBinding(sqlDigests []string) (deletedRows uint64, err error) {
-	if len(sqlDigests) == 0 {
-		return 0, errors.New("sql digest is empty")
-	}
-	defer func() {
-		if err == nil {
-			err = h.LoadFromStorageToCache(false)
-		}
-	}()
-
-	err = callWithSCtx(h.sPool, true, func(sctx sessionctx.Context) error {
-		// Lock mysql.bind_info to synchronize with CreateBinding / AddBinding / DropBinding on other tidb instances.
-		if err = lockBindInfoTable(sctx); err != nil {
-			return err
-		}
-
-		for _, sqlDigest := range sqlDigests {
-			updateTs := types.NewTime(types.FromGoTime(time.Now()), mysql.TypeTimestamp, 3).String()
-			_, err = exec(
-				sctx,
-				`UPDATE mysql.bind_info SET status = %?, update_time = %? WHERE sql_digest = %? AND update_time < %? AND status != %?`,
-				StatusDeleted,
-				updateTs,
-				sqlDigest,
-				updateTs,
-				StatusDeleted,
-			)
-			if err != nil {
-				return err
-			}
-			deletedRows += sctx.GetSessionVars().StmtCtx.AffectedRows()
-		}
-		return nil
-	})
-	if err != nil {
-		deletedRows = 0
-	}
-	return deletedRows, err
-}
-
-// SetGlobalBindingStatus set a Bindings's status to the storage and bind cache.
-func (h *globalBindingHandle) SetGlobalBindingStatus(newStatus, sqlDigest string) (ok bool, err error) {
-	var (
-		updateTs               types.Time
-		oldStatus0, oldStatus1 string
-	)
-	if newStatus == StatusDisabled {
-		// For compatibility reasons, when we need to 'set binding disabled for <stmt>',
-		// we need to consider both the 'enabled' and 'using' status.
-		oldStatus0 = StatusUsing
-		oldStatus1 = StatusEnabled
-	} else if newStatus == StatusEnabled {
-		// In order to unify the code, two identical old statuses are set.
-		oldStatus0 = StatusDisabled
-		oldStatus1 = StatusDisabled
-	}
-
-	defer func() {
-		if err == nil {
-			err = h.LoadFromStorageToCache(false)
-		}
-	}()
-
-	err = callWithSCtx(h.sPool, true, func(sctx sessionctx.Context) error {
-		// Lock mysql.bind_info to synchronize with SetBindingStatus on other tidb instances.
-		if err = lockBindInfoTable(sctx); err != nil {
-			return err
-		}
-
-		updateTs = types.NewTime(types.FromGoTime(time.Now()), mysql.TypeTimestamp, 3)
-		updateTsStr := updateTs.String()
-
-		_, err = exec(sctx, `UPDATE mysql.bind_info SET status = %?, update_time = %? WHERE sql_digest = %? AND update_time < %? AND status IN (%?, %?)`,
-			newStatus, updateTsStr, sqlDigest, updateTsStr, oldStatus0, oldStatus1)
-		return err
-	})
-	return
-}
-
-// GCGlobalBinding physically removes the deleted bind records in mysql.bind_info.
-func (h *globalBindingHandle) GCGlobalBinding() (err error) {
-	return callWithSCtx(h.sPool, true, func(sctx sessionctx.Context) error {
-		// Lock mysql.bind_info to synchronize with CreateBinding / AddBinding / DropBinding on other tidb instances.
-		if err = lockBindInfoTable(sctx); err != nil {
-			return err
-		}
-
-		// To make sure that all the deleted bind records have been acknowledged to all tidb,
-		// we only garbage collect those records with update_time before 10 leases.
-		updateTime := time.Now().Add(-(10 * Lease))
-		updateTimeStr := types.NewTime(types.FromGoTime(updateTime), mysql.TypeTimestamp, 3).String()
-		_, err = exec(sctx, `DELETE FROM mysql.bind_info WHERE status = 'deleted' and update_time < %?`, updateTimeStr)
-		return err
-	})
 }
 
 // lockBindInfoTable simulates `LOCK TABLE mysql.bind_info WRITE` by acquiring a pessimistic lock on a
@@ -363,10 +183,4 @@ func (h *globalBindingHandle) Stats(_ *variable.SessionVars) (map[string]any, er
 	m := make(map[string]any)
 	m[lastPlanBindingUpdateTime] = h.LastUpdateTime().String()
 	return m, nil
-}
-
-// Close closes the binding handler.
-func (h *globalBindingHandle) Close() {
-	h.BindingCacheUpdater.Close()
-	h.sPool.Close()
 }
